Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
 int main(){
-    float a,b;
-    scanf("%f%f",&a,&b);
-    if(a>=b){printf("True");}
+    int a,b;
+    scanf("%d%d",&a,&b);
+    if(a>=b){
+        printf("True");}
     else{
         printf("False");
     }
